+++ server/src/main/java/org/elasticsearch/common/bytes/BytesArray.java
@@ -21,7 +21,7 @@ package org.elasticsearch.common.bytes;
 
 import org.apache.lucene.util.BytesRef;
 
-public final class BytesArray extends BytesReference {
+public final class BytesArray extends AbstractBytesReference {
 
     public static final BytesArray EMPTY = new BytesArray(BytesRef.EMPTY_BYTES, 0, 0);
     private final byte[] bytes;
