+++ server/src/main/java/org/elasticsearch/common/bytes/PagedBytesReference.java
@@ -30,7 +30,7 @@ import java.io.IOException;
  * A page based bytes reference, internally holding the bytes in a paged
  * data structure.
  */
-public class PagedBytesReference extends BytesReference {
+public class PagedBytesReference extends AbstractBytesReference {
 
     private static final int PAGE_SIZE = PageCacheRecycler.BYTE_PAGE_SIZE;
