+++ server/src/main/java/org/elasticsearch/common/io/stream/ReleasableBytesStreamOutput.java
@@ -19,7 +19,8 @@
 
 package org.elasticsearch.common.io.stream;
 
-import org.elasticsearch.common.bytes.ReleasablePagedBytesReference;
+import org.elasticsearch.common.bytes.PagedBytesReference;
+import org.elasticsearch.common.bytes.ReleasableBytesReference;
 import org.elasticsearch.common.lease.Releasable;
 import org.elasticsearch.common.lease.Releasables;
 import org.elasticsearch.common.util.BigArrays;
@@ -31,7 +32,7 @@ import org.elasticsearch.common.util.PageCacheRecycler;
  * expecting it to require releasing its content ({@link #bytes()}) once done.
  * <p>
  * Please note, closing this stream will release the bytes that are in use by any
- * {@link ReleasablePagedBytesReference} returned from {@link #bytes()}, so this
+ * {@link ReleasableBytesReference} returned from {@link #bytes()}, so this
  * stream should only be closed after the bytes have been output or copied
  * elsewhere.
  */
@@ -55,8 +56,8 @@ public class ReleasableBytesStreamOutput extends BytesStreamOutput
      * the bytes in the stream.
      */
     @Override
-    public ReleasablePagedBytesReference bytes() {
-        return new ReleasablePagedBytesReference(bytes, count, releasable);
+    public ReleasableBytesReference bytes() {
+        return new ReleasableBytesReference(new PagedBytesReference(bytes, count), releasable);
     }
 
     @Override
